frame: Re-work frame decoding.
Put the majority of the frame logic in the top-level `AX25Frame.decode`
method for convenience.

We can rely on the fact that the least-significant two bits of a control
field are all we need to identify I, S and U frames.  Further more, the
least significant byte is sent first, so there's sufficient detail at
this point to identify the U frames up-front, and with some help,
I and S frames too.